requirements.txt
openai
langchain
faiss-cpu
chromadb
pyyaml
gitpython
ghapi
biopython
requests

# Jupyter and dashboard dependencies
jupyter
notebook
jupyterlab
pandas
matplotlib
plotly
ipywidgets